requirements.txt
azure-common==1.1.26
azure-core==1.10.0
azure-mgmt-core==1.2.2
azure-mgmt-datafactory==1.0.0
azure-mgmt-storage==16.0.0
azure-storage-blob==12.7.0
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
idna==2.10
isodate==0.6.0
msrest==0.6.19
numpy==1.19.5
oauthlib==3.1.0
pandas==1.2.0
paramiko==2.7.2
pycparser==2.20
PyNaCl==1.4.0
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
storage==0.0.4.3
urllib3==1.26.2
